// types.ts
export type Permissions = Record<string, boolean>;

export interface Role {
    id: string;
    name: string;
    permissions: Permissions;
}

export interface User {
    id: string;
    name: string; // Full display name
    username: string; // Login username
    email?: string; // Optional email
    roleId: string;
    officeId?: string;
    password?: string;
}

export interface CompanyInfo {
    name: string;
    rif: string;
    address: string;
    phone: string;
    logoUrl?: string;
    loginImageUrl?: string;
    costPerKg?: number;
    bcvRate?: number;
    postalLicense?: string;
}

export interface Office {
    id: string;
    code: string;
    name: string;
    address: string;
    phone: string;
}
// ...
export interface Category {
    id: string;
    name: string;
}

export interface Product {
    id: string;
    name: string;
    description?: string;
}

export interface Client {
    id: string;
    idNumber: string; // RIF/Cédula
    clientType: 'persona' | 'empresa';
    name: string;
    phone: string;
    address: string;
    email?: string;
}

export interface Supplier {
    id: string;
    idNumber: string; // RIF
    name: string; // Razón Social
    phone: string;
    address: string;
}

export interface Merchandise {
    quantity: number;
    weight: number;
    length: number;
    width: number;
    height: number;
    description: string;
    categoryId: string;
}

export interface ShippingGuide {
    guideNumber: string;
    date: string;
    originOfficeId: string;
    destinationOfficeId: string;
    sender: Partial<Client>;
    receiver: Partial<Client>;
    merchandise: Merchandise[];
    shippingTypeId: string;
    paymentMethodId: string;
    hasInsurance: boolean;
    declaredValue: number;
    insurancePercentage: number;
    paymentType: 'flete-pagado' | 'flete-destino';
    paymentCurrency: 'VES' | 'USD';
    hasDiscount: boolean;
    discountPercentage: number;
    // New fields
    pickupOrder?: string;
    isTransbordo?: boolean;
    receivedBy?: string;
}

export interface Financials {
    freight: number;
    insuranceCost: number;
    handling: number;
    discount: number;
    subtotal: number;
    ipostel: number;
    iva: number;
    igtf: number;
    total: number;
}

export type PaymentStatus = 'Pagada' | 'Pendiente';
export type ShippingStatus = 'Pendiente para Despacho' | 'En Tránsito' | 'Entregada';
export type MasterStatus = 'Activa' | 'Anulada';

export interface Invoice {
    id: string;
    invoiceNumber: string;
    controlNumber: string;
    date: string;
    clientName: string;
    clientIdNumber: string;
    totalAmount: number;
    status: MasterStatus;
    paymentStatus: PaymentStatus;
    shippingStatus: ShippingStatus;
    guide: ShippingGuide;
    vehicleId?: string;
    remesaId?: string;
    createdByName?: string; // To store "Oficinista" name
}